chore: log extra args through auto-instrumentation
